Fix gcc warning on main
The K&R book was written when the main function did not need to be
declared as a type. Some systems defaulted to void and others did to
int. In C standards past 1977, it is seen as good practice to declare
it explicitely as int if you are going to return a value from it for
error or success.

Added in stdlib and exit(0) at the end as also good behaviour.