Update the code in s1.5.2_count to match second example on pg 18
I don't like the second example as it shows a lot of shortcuts in C
which tend to bite people over time when trying to manage other
people's code.  First everything is being done in the `for` statement
which can lead to needing to rewrite the code completely on small
changes.