Fix warnings from splint.
As part of the finger exercises to get into a better programmer, I am
using various lint tools. Splint showed that I was comparing different
types and also not looking at the return from certain functions. Since
that could lead to issues depending on the system, I was more explicit
in what was wanted.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>