Import example s01.05.04 from K&R pg 20
This is another version of the 'counting' character program in
K&R. This version is closer to the wc program in counting some items
and outputting basic characters, words, and lines in a file. However
it does not have any logic to open the file or limit what output is to
be given.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>