Fix splint warnings in wc program.
Like many early C programs, there is a fast and loose comparison
between int and char characters. This can and will cause problems as
char is limited to 8 bits and specific 'characters' while int is
architecture dependent and can be 8, 16, 32, 64 bit in size.

Splint warns on this and the 'fix' in general is to cast the int
variable to a char. This also has side-effects in that the int may not
be 8 bits in size but some large number.

The reason for this failure of types is that the 'getchar' function
does not return a char but an int. The thudding sound you are hearing
is usually from new programmers to the C language hitting their head
against the table. Don't worry folks, it only gets worse from here.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>