Complete exercise 12
Program to take input of text, and then print one per line ignoring
blanks and newlines.

Wrote some notes to cover exercise 11. Would be better to add some
script to run the tests.

Rework the Makefile a little bit to make it a little less
awkward. Needs more work as it is getting long and hard to cover.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>