Input sample from K&R C, pg22 S01.06
This program gives an example array and then uses some counting
methods to count the number of each digit in a file. The output needs
work as it does not cover say what digit is being counted where. I also deviated in this initial checkin from the original text to cover some warnings from splint on char/int math.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>