Clean up some warnings from splint on histogram program.
The usual warnings came up about not looking at the return of putchar
and also that '\n' is not an int. Using casting stopped those. There
is an outstanding warning that initializing an array with {0} only
does one value so I am guessing I am using a gcc code here.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>