Complete K&R exercise 14
This program will do a histogram of ascii characters and will try to
keep track of characters outside of the 'ascii' range seen. In writing
this I made the mistake of sending to isalpha() the value inside the
array at a counter versus the counter. This caused a segmentation
fault but I did not see this listed as a known error in the man
page. I expect it is a undefined response somewhere.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>