Make a minimally updated version of K&R C pg 29 program
This program is meant to be an example of how to use arrays and
functions and th ability to pass arrays by address versus by value. It
has a lot of problems in the shortcuts the program used as written. I
have changed the while{} to a do{} since it is rarely used and I don't
like having so many side effect calls in a test.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>