Complete K&R exercise 01.16
This one was mainly an exercise in changing the existing program to do
two things: One print the longest line found and make sure that if the
line was longer than could be stored that it was cleaned up
correctly. This meant making x[k-1] to be '\0' and x[k-2] is '\n'.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>