Worked through K&R exercise 18
In this case, I took code from exercise 17 and used it to getlines. I
then added a routine to go from the known length of line code to
figure out where the end of line was and work backwards. A function by
itself would have needed to walk the array until it either found a \0
or \n and then walk backwards.