Work mainly on K&R exercise 19 reverse program
Saw that I did not clean a comment of code in ex18 which I had gotten
from ex19.

Worked on Exercise 18 which is to reverse a string. Most of the work
is using a pretty standard coding function which reverses an array. It
does this by finding the end of a string for one variable and then
walks both ends and swaps the items.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>