Added example for Exercise 23 of K&R Chapter 1
This example starts having programs which can act like steps of a
compiler. In this case, we are removing all the comments from some
code which might make something to strip out before the next steps of
a compiler works.

Things I learned in this is that you need to account for both ' and "
quotes. Just looking at " had the system going until the end of time
not removing comments for some reason.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>