K&R Chapter 2, problem 2-3. setbits
This problem took me a while to figure it out and I ended up
having to go to the C Answer Book in order to get past things.
That said, the one liner they showed in the book really needs a
full explantion in code which I tried to put together for
future me. [I have been making mistakes with bit operations for
a long time.]

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>